// interfaces/ITrendingMovies.ts
export interface ITrendingMovies {
    page:          number;
    results:       ITrendingMovie[];
    total_pages:   number;
    total_results: number;
}

export interface ITrendingMovie {
    backdrop_path:     string;
    id:                number;
    title:             string;
    original_title:    string;
    overview:          string;
    poster_path:       string;
    media_type:        MediaType;
    adult:             boolean;
    original_language: string;
    genre_ids:         number[];
    popularity:        number;
    release_date:      Date;
    video:             boolean;
    vote_average:      number;
    vote_count:        number;
}

export enum MediaType {
    Movie = "movie",
}